extract the input detection into a separate hook so it's reusable
