+++ studio/src/pages/RequestDetailsPage/hooks/index.ts
@@ -3,3 +3,4 @@ export { usePagination } from "./usePagination";
 export { useReplayRequest } from "./useReplayRequest";
 export { useMostRecentRequest } from "./useMostRecentRequest";
 export { useShouldReplay } from "./useShouldReplay";
+export { useInputFocusDetection } from "./useInputFocusDetection";
