+++ studio/src/pages/RequestDetailsPage/hooks/useNavigateToList.ts
@@ -1,45 +1,19 @@
-import { useEffect, useState } from "react";
 import { useHotkeys } from "react-hotkeys-hook";
 import { useNavigate } from "react-router-dom";
+import { useInputFocusDetection } from "./useInputFocusDetection";
 
 export function useEscapeToList() {
   const navigate = useNavigate();
-
-  const [isInputFocused, setIsInputFocused] = useState(false);
+  const { isInputFocused, blurActiveInput } = useInputFocusDetection();
 
   useHotkeys(["Escape"], () => {
     // catch all the cases where the user is in the input field
     // and we don't want to exit the page
     if (isInputFocused) {
-      const activeElement = document.activeElement;
-      if (activeElement instanceof HTMLInputElement) {
-        activeElement.blur();
-      }
+      blurActiveInput();
       return;
     }
 
     navigate("/requests");
   });
-
-  useEffect(() => {
-    const handleFocus = (event: FocusEvent) => {
-      if (event.target instanceof HTMLInputElement) {
-        setIsInputFocused(true);
-      }
-    };
-    const handleBlur = (event: FocusEvent) => {
-      if (event.target instanceof HTMLInputElement) {
-        setIsInputFocused(false);
-      }
-    };
-
-    // We can use AbortController to remove both event listeners a bit more cleanly
-    // https://frontendmasters.com/blog/patterns-for-memory-efficient-dom-manipulation/#use-abortcontroller-to-unbind-groups-of-events
-    document.addEventListener("focus", handleFocus, true);
-    document.addEventListener("blur", handleBlur, true);
-    return () => {
-      document.removeEventListener("focus", handleFocus, true);
-      document.removeEventListener("blur", handleBlur, true);
-    };
-  }, []);
 }
